refactored registry auxiliary code for e2e tests
